+++ Integrations/Elasticsearch_v2/Elasticsearch_v2_test.py
@@ -403,17 +403,3 @@ def test_incident_creation_with_timestamp_e7():
     incidents, last_fetch2 = results_to_incidents_timestamp(ES_V7_RESPONSE_WITH_TIMESTAMP, lastfetch)
     assert last_fetch2 == 1572502640
     assert str(incidents) == MOCK_ES7_INCIDENTS_FROM_TIMESTAMP
-
-
-@patch("Elasticsearch_v2.TIME_METHOD", 'Simple-Date')
-@patch("Elasticsearch_v2.TIME_FIELD", 'Date')
-@patch("Elasticsearch_v2.FETCH_TIME", "5 days")
-@patch("Elasticsearch_v2.TIME_FORMAT", '%Y-%m-%dT%H:%M:%SZ')
-@patch("Elasticsearch_v2.SERVER", 'http://localhost:9200')
-def test_fetch_indicators_command():
-    from Elasticsearch_v2 import fetch_indicators_command
-    fetch_indicators_command()
-    # lastfetch = int(datetime.strptime('2019-08-27T17:59:00Z', '%Y-%m-%dT%H:%M:%SZ').timestamp())
-    # incidents, last_fetch2 = results_to_incidents_timestamp(ES_V7_RESPONSE_WITH_TIMESTAMP, lastfetch)
-    # assert last_fetch2 == 1572502640
-    # assert str(incidents) == MOCK_ES7_INCIDENTS_FROM_TIMESTAMP