+++ Integrations/Elasticsearch_v2_5_5/Elasticsearch_v2_5_5.py
@@ -17,8 +17,13 @@ requests.packages.urllib3.disable_warnings()
 warnings.filterwarnings(action="ignore", message='.*using SSL with verify_certs=False is insecure.')
 
 SERVER = demisto.params().get('url', '').rstrip('/')
-USERNAME = demisto.params().get('credentials', {}).get('identifier')
-PASSWORD = demisto.params().get('credentials', {}).get('password')
+CREDS = demisto.params().get('credentials')
+if CREDS:
+    USERNAME = CREDS.get('identifier')
+    PASSWORD = CREDS.get('password')
+else:
+    USERNAME = None
+    PASSWORD = None
 PROXY = demisto.params().get('proxy')
 HTTP_ERRORS = {
     400: '400 Bad Request - Incorrect or invalid parameters',
@@ -31,7 +36,6 @@ HTTP_ERRORS = {
     503: '503 Service Unavailable'
 }
 
-
 '''VARIABLES FOR FETCH INCIDENTS'''
 TIME_FIELD = demisto.params().get('fetch_time_field', '')
 TIME_FORMAT = demisto.params().get('fetch_time_format', '')
@@ -41,6 +45,7 @@ FETCH_TIME = demisto.params().get('fetch_time', '3 days')
 FETCH_SIZE = int(demisto.params().get('fetch_size', 50))
 INSECURE = not demisto.params().get('insecure', False)
 TIME_METHOD = demisto.params().get('time_method', 'Simple-Date')
+MODULE_TO_FEEDMAP_KEY = 'moduleToFeedMap'
 
 # if timestamp than set the format to iso.
 if 'Timestamp' in TIME_METHOD:
@@ -509,10 +514,71 @@ def results_to_incidents_datetime(response, last_fetch):
     return incidents, last_fetch
 
 
-def fetch_incidents():
+def get_indicators_command():
+    search, _ = get_indicators_search_scan()
+    limit = int(demisto.args().get('limit', FETCH_SIZE))
+    indicators_list: list = []
+    for hit in search.scan():
+        indicators_list.extend(extract_indicators_from_insight_hit(hit))
+        if len(indicators_list) >= limit:
+            break
+    hr = tableToMarkdown('Indicators', indicators_list, ['name'])
+    return_outputs(hr, {'ElasticsearchFeed.SharedIndicators': indicators_list}, indicators_list)
+
+
+def fetch_indicators_command():
+    search, now = get_indicators_search_scan()
+    ioc_lst: list = []
+    for hit in search.scan():
+        ioc_lst.extend(extract_indicators_from_insight_hit(hit))
+    if ioc_lst:
+        for b in batch(ioc_lst, batch_size=2000):
+            demisto.createIndicators(b)
+    demisto.setLastRun({'time': now})
+
+
+def get_indicators_search_scan():
+    now = datetime.now()
+    time_field = "calculatedTime"
+    last_fetch = demisto.getLastRun().get('time')
+    range_field = {time_field: {'gt': datetime.fromtimestamp(float(last_fetch)), 'lte': now}} if last_fetch else {
+        time_field: {'lte': now}}
+    es = elasticsearch_builder()
+    query = QueryString(query=time_field + ":*")
+    tenant_hash = demisto.getIndexHash()
+    # all shared indexes minus this tenant shared
+    indexes = f'*-shared*,-*{tenant_hash}*-shared*'
+    search = Search(using=es, index=indexes).filter({'range': range_field}).query(query)
+    return search, str(now.timestamp())
+
+
+def extract_indicators_from_insight_hit(hit):
+    ioc_lst = []
+    ioc = results_to_indicator(hit)
+    if ioc.get('value'):
+        ioc_lst.append(ioc)
+        module_to_feedmap = ioc.get(MODULE_TO_FEEDMAP_KEY)
+        updated_module_to_feedmap = {}
+        if module_to_feedmap:
+            for key, val in module_to_feedmap.items():
+                if val.get('isEnrichment'):
+                    ioc_lst.append(val)
+                else:
+                    updated_module_to_feedmap[key] = val
+            ioc[MODULE_TO_FEEDMAP_KEY] = updated_module_to_feedmap
+    return ioc_lst
+
+
+def results_to_indicator(hit):
+    ioc_dict = hit.to_dict()
+    ioc_dict['value'] = ioc_dict.get('name')
+    ioc_dict['rawJSON'] = dict(ioc_dict)
+    return ioc_dict
+
+
+def get_last_fetch_time():
     last_run = demisto.getLastRun()
     last_fetch = last_run.get('time')
-
     # handle first time fetch
     if last_fetch is None:
         last_fetch, _ = parse_date_range(date_range=FETCH_TIME, date_format=TIME_FORMAT, utc=False, to_timestamp=False)
@@ -525,7 +591,11 @@ def fetch_incidents():
     # if method is simple date - convert the date string to datetime
     elif 'Simple-Date' == TIME_METHOD:
         last_fetch = datetime.strptime(last_fetch, TIME_FORMAT)
+    return last_fetch
 
+
+def fetch_incidents():
+    last_fetch = get_last_fetch_time()
     es = elasticsearch_builder()
 
     query = QueryString(query=FETCH_QUERY + " AND " + TIME_FIELD + ":*")
@@ -557,6 +627,10 @@ def main():
             test_func()
         elif demisto.command() == 'fetch-incidents':
             fetch_incidents()
+        elif demisto.command() == 'fetch-indicators':
+            fetch_indicators_command()
+        elif demisto.command() == 'get-shared-indicators':
+            get_indicators_command()
         elif demisto.command() in ['search', 'es-search']:
             search_command()
     except Exception as e: