+++ Integrations/Elasticsearch_v2/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
-- Added ***fetch-indicators*** support, intended to be used to share IoCs between tenants in an Elasticsearch MT environment.
-- Added ***get-shared-indicators*** command used to test ***fetch-indicators***.
+  - Added support for the ***fetch-indicators*** function, which is used to share IoCs between tenants in multi-tenant deployments.
+  - Added the ***get-shared-indicators*** command, which tests the ***fetch-indicators*** function.
 
 ## [19.12.0] - 2019-12-10
 -