+++ Integrations/Elasticsearch_v2/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## [Unreleased]
 
+
+## [19.12.0] - 2019-12-10	
+-	
+
 ## [19.11.0] - 2019-11-12
-Added support for timestamps. 
+Added support for timestamps.
     
 ## [19.9.1] - 2019-09-18
 New Elasticsearch v2 integration.